The `Obj'` type family now depends on object as well as category.
The `Category` class in section 6 of [*Type Your Matrices*](https://github.com/bolt12/tymfgg-pearl) uses an associated constraint that depends on the object as well as the category.
Mine depended only on the category and so sometimes necessitates defining new classes.
I carried over this choice from concat, but I guess the motivation has gone away with the introduction of quantified constraints or some other change.
(Now I wonder if concat even needs it.
In either case, I should probably switch concat over to quantified constraints as well and eliminate all of the constraint entailment programming.)
I suspect it helps considerably that we define an [extra class that wraps the associated `Obj'` constraint](https://github.com/conal/linalg/pull/28#issuecomment-670313952).

This change eliminated the need for `Unconstrained`, so I dropped it.